+++ libSyncUpdate/sync_server/sync_server_main.cpp
@@ -316,7 +316,7 @@ int sync_server_cmd_line(int argc, const char * argv[]){
     int hashClashBit=estimateHashClashBit(newDataSize,(uint32_t)kMatchBlockSize);
     _return_check(hashClashBit<=kAllowMaxHashClashBit,
                   SYNC_SERVER_BLOCKSIZE_ERROR,"hash clash warning! must increase matchBlockSize(%d) !\n",(uint32_t)kMatchBlockSize);
-    double patchMemSize=estimatePatchMemSize(newDataSize,(uint32_t)kMatchBlockSize,(compressPlugin!=0));
+    double patchMemSize=(double)estimatePatchMemSize(newDataSize,(uint32_t)kMatchBlockSize,(compressPlugin!=0));
     if (patchMemSize>=(1<<20))
         printf("sync_patch memory size: ~ %.3f MB\n",patchMemSize/(1<<20));
     else
